Cache improvements, configuration, and docs (#135)
Added the ability to configure caches through a configuration file.

Improved cache behaviour when multiple caches are present and now allow
configuration of which caches are active. Added determinism checking for
when objects have the same transform hash but different output hashes.

Added documentation on caches.

Documented further improvements in #136